Added functionality for user to specify blindness treetree
Combined the two findOptimalWitness functions into one that should work for general N.
Defined new getBlindness function that returns a "blindfold" that can be elementwise multiplied with the witness to produce the blindness conditions. Also outputs a schematic of the blindness tree.
Modified  findggme to work with new function (removed some ifs).
Modified produceggme to include a line for getBlindness and an earlier setting of  only_partial_knowlege